MOVE

+++ GameEngine/include/RType.hpp
@@ -40,8 +40,8 @@ namespace rtype
     struct Entity
     {
         uint16_t id; // client
-        uint32_t positionX;
-        uint32_t positionY;
+        float positionX;
+        float positionY;
         float directionX;
         float directionY;
     };
